requirements.txt
django==5.1.2
djangorestframework==3.15.2
request==2.32.3